fix: auto-create bucket on first share, add missing AppHeader component
